Refactor server - separate action/loader handlers for node, add reader api for local mode
