--- packages/server/src/index.ts
@@ -1,2 +1,0 @@
-export * from '~/handlers/actions'
-export * from '~/handlers/loaders'
