+++ packages/server/src/types.ts
@@ -1,4 +1,4 @@
-import type { Config } from '@kobun/common'
+import type { Config, ConfigSchema, SchemaKey } from '@kobun/common'
 import type { ActionFunctionArgs, LoaderFunctionArgs } from 'react-router'
 
 export type ActionHandlerArgs = {
@@ -8,3 +8,34 @@ export type ActionHandlerArgs = {
 export type LoaderHandlerArgs = {
 	config: Config
 } & LoaderFunctionArgs
+
+export type AllParams = {
+	filters?: Record<string, string>
+}
+
+export type UniqueArg = {
+	where: {
+		id?: string
+		slug?: string
+	}
+	options?: {
+		headings: boolean
+	}
+}
+
+export type UniqueReturn = {
+	code: string
+	content: string
+	frontmatter: unknown
+}
+
+export type CollectionInterface<T extends SchemaKey> = {
+	_label: string
+	_schema: ConfigSchema<T>
+	all: (params: AllParams) => Promise<unknown[]>
+	unique: (arg: UniqueArg) => Promise<unknown>
+}
+
+export type KobunReader<T extends SchemaKey> = {
+	[key: string]: CollectionInterface<T>
+}
